# internal/web/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: qui API
  description: qui API for managing multiple qBittorrent instances
  version: 1.0.0
  license:
    name: GPL-2.0-or-later
    url: https://opensource.org/licenses/GPL-2.0
servers:
  - url: http://localhost:7476
    description: Local development server
  - url: /
    description: Current server
security:
  - ApiKeyAuth: []
  - SessionAuth: []
paths:
  /api/auth/setup:
    post:
      tags:
        - Authentication
      summary: Initial setup
      description: Create the initial admin user (only available before first user is created)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 8
                  writeOnly: true
                  description: Password will be hashed and never returned in responses
      responses:
        '201':
          description: Setup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Setup already completed
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    SessionAuth:
      type: apiKey
      in: cookie
      name: user_session
      description: Session cookie authentication
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
...